+++ spec/controllers/controller_spec.rb
@@ -52,7 +52,7 @@ describe SorceryController, type: :controller do
     describe '#login' do
       context 'when succeeds' do
         before do
-          expect(User).to receive(:authenticate).with('bla@bla.com', 'secret').and_return(user)
+          expect(User).to receive(:authenticate).with('bla@bla.com', 'secret') { |&block| block.call(user, nil) }
           get :test_login, params: { email: 'bla@bla.com', password: 'secret' }
         end
 
@@ -132,7 +132,7 @@ describe SorceryController, type: :controller do
       sorcery_controller_property_set(:not_authenticated_action, :test_not_authenticated_action)
       get :test_logout
 
-      expect(response).to be_a_success
+      expect(response).to be_successful
     end
 
     it 'require_login before_action saves the url that the user originally wanted' do